Move slurm settings under expandable section and add placeholder fields
